Overhaul materials system
Move material properties from uniforms to vertex attributes to allow for mesh flattening in the future; choose whether to include normals in the WebGL mesh based on what kind of material is selected (normals are not needed for flat or emissive materials). Also switch from Vec3 to Color for flat materials to emphasize that no lighting/gamma correction is performed.