Tweak cone normals for smoother shading
Support zero-length normals in vertex shader so cone tip can be given a zero normal; this means that even very close to the tip the normal will interpolate properly from the 'left' normal to the 'right' normal. With the tip normal set to the average of the left and right normals, the interpolated normal ends up being 'flattened' near the tip.